requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
pyrebase4
firebase-admin
